extract LineFeeder to its own class, with tests
