+++ src/zork/freedom-module.ts
@@ -7,6 +7,7 @@ import churn_types = require('../churn/churn.types');
 import logging = require('../logging/logging');
 import loggingTypes = require('../loggingprovider/loggingprovider.types');
 import net = require('../net/net.types');
+import linefeeder = require('../net/linefeeder');
 import queue = require('../handler/queue');
 import rtc_to_net = require('../rtc-to-net/rtc-to-net');
 import socks_to_rtc = require('../socks-to-rtc/socks-to-rtc');
@@ -66,29 +67,6 @@ var sendReply = (message:string, connection:tcp.Connection) : void => {
   connection.send(arraybuffers.stringToArrayBuffer(message + '\n'));
 }
 
-// Transforms a raw network-like ArrayBuffer-based queue into
-// a telnet-style string-based queue.
-class LineFeeder extends queue.Queue<string, void> {
-  private static DELIMITER = arraybuffers.decodeByte(
-      arraybuffers.stringToArrayBuffer('\n'));
-
-  constructor(source: queue.Queue<ArrayBuffer, void>) {
-    super();
-    let leftover = new ArrayBuffer(0);
-    source.setSyncHandler((buffer: ArrayBuffer) => {
-      leftover = arraybuffers.concat([leftover, buffer]);
-      let i = arraybuffers.indexOf(leftover, LineFeeder.DELIMITER);
-      while (i !== -1) {
-        let parts = arraybuffers.split(leftover, i);
-        let line = arraybuffers.arrayBufferToString(parts[0]);
-        leftover = parts[1].slice(1);
-        i = arraybuffers.indexOf(leftover, LineFeeder.DELIMITER);
-        this.handle(line);
-      }
-    });
-  }
-}
-
 // Rough lifecycle is to process single word commands such as "ping" until
 // "get" or "give" is received, at which point the connection is handed off
 // to a SocksTotc or RtcToNet instance (and further input is treated as
@@ -97,7 +75,7 @@ function serveConnection(connection: tcp.Connection): void {
   var transformerName: string;
   var transformerConfig: string;
 
-  const lineFeeder = new LineFeeder(connection.dataFromSocketQueue);
+  const lineFeeder = new linefeeder.LineFeeder(connection.dataFromSocketQueue);
   var doit = (command: string) => {
     log.debug('%1: received command: %2', connection.connectionId, command);
